+++ src/index.ts
@@ -4,6 +4,7 @@ export * from "./dnxDecoder.js";
 export * from "./dnxFrame.js";
 export * from "./dnxMxf.js";
 export * from "./dnxPlaybackClock.js";
+export { deinterlaceDnxFrameLayout, type DnxFieldParity } from "./dnxPixelConversion.js";
 export * from "./dnxMediabunny.js";
 export * from "./dnxRandomAccessDecoder.js";
 export * from "./dnxWebGpuRenderer.js";